Substitui variaveis do tipo Double para BigDecimal

+++ src/main/java/br/me/desafio/challengeme/DTO/PedidoItemRespostaDTO.java
@@ -1,16 +1,18 @@
 package br.me.desafio.challengeme.DTO;
 
+import java.math.BigDecimal;
+
 public class PedidoItemRespostaDTO {
 
     private String descricao;
-    private Double precoUnitario;
+    private BigDecimal precoUnitario;
     private Integer qtd;
 
     public String getDescricao() {
         return descricao;
     }
 
-    public Double getPrecoUnitario() {
+    public BigDecimal getPrecoUnitario() {
         return precoUnitario;
     }
 
@@ -18,7 +20,7 @@ public class PedidoItemRespostaDTO {
         return qtd;
     }
 
-    public PedidoItemRespostaDTO(String descricao, Double precoUnitario, Integer qtd) {
+    public PedidoItemRespostaDTO(String descricao, BigDecimal precoUnitario, Integer qtd) {
         this.descricao = descricao;
         this.precoUnitario = precoUnitario;
         this.qtd = qtd;
